requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.3
brotli==1.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-daisyui==0.7.0
cryptography==45.0.4
cssselect2==0.8.0
dj-database-url==2.3.0
django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-htmx==1.23.0
django-jazzmin==3.0.1
django-parler==2.3
django-parler-rest==2.2
django-segno-qr==1.0.2
djangorestframework==3.16.0
fonttools==4.58.0
gitdb==4.0.12
gitpython==3.1.44
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
markdown==3.8
nanoid==2.0.0
nh3==0.2.21
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
prompt-toolkit==3.0.52
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydyf==0.11.0
pyjwt==2.10.1
pyphen==0.17.2
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-frontmatter==1.1.0
pyyaml==5.1
redis==7.1.0
requests==2.32.4
ruff==0.11.11
segno==1.6.6
sentry-sdk==2.29.1
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
